Backend/complistrux/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
docker==6.0.1
Faker==15.3.4
idna==3.4
lxml==4.9.2
numpy==1.23.4
packaging==23.0
pandas==1.5.1
Pillow==9.4.0
psycopg2==2.9.5
pyaml==21.10.1
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
pytz==2022.6
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
websocket-client==1.4.2